+++ packages/themes/Styrke/src/export.js
@@ -12,14 +12,15 @@ import {
 import { TextStyling } from "./Theme/theme.Texts"
 import { WrapperStyling } from "./Theme/theme.Wrappers"
 import { ItemGroupStyling } from "./Theme/theme.ItemGroups"
+import { RadioPillItemStyling } from "./Theme/theme.RadioPillItems"
+import { SelectStyling } from "./Theme/theme.Selects"
 import { ButtonStyling } from "./Theme/theme.Buttons"
 import {
   InputStyling,
   LabelStyling,
   SelectSelectedStyling,
   SelectIconButtonStyling,
-  SelectOptionsWrapperStyling,
-  RadioPillStyling
+  SelectOptionsWrapperStyling
 } from "./Theme/theme.Inputs"
 import { LayoutStyling, LayoutItemStyling } from "./Theme/theme.Layouts"
 import { ExpandStyling } from "./Theme/theme.Expand"
@@ -66,9 +67,9 @@ const color = {
   yellow: "#FFF8BB",
   line: "#d8d8d8",
   wcag: "#777",
-  disabled: "#BABABA",
+  disabled: "#a0a0a0",
   warning: "#E06666",
-  black: "#000",
+  black: "#333",
   subtleHover: "#fff"
 }
 
@@ -123,6 +124,8 @@ export default new theming.Theme({
   .add(TextStyling)
   .add(WrapperStyling)
   .add(ItemGroupStyling)
+  .add(RadioPillItemStyling)
+  .add(SelectStyling)
   .add(ButtonStyling)
   .add(SplitListItemStyling)
   .add(InputStyling)
@@ -133,6 +136,5 @@ export default new theming.Theme({
   .add(SelectOptionsWrapperStyling)
   .add(LayoutStyling)
   .add(LayoutItemStyling)
-  .add(RadioPillStyling)
   .add(ExpandStyling)
   .add(DividerStyling)